+++ meta/template/table.lua
@@ -57,11 +57,22 @@ function table.sort(list, comp) end
 
 ---@version >5.2, JIT
 ---#DES 'table.unpack'
----@generic T
----@param list T[]
+---@generic T1, T2, T3, T4, T5, T6, T7, T8, T9, T10
+---@param list {
+--- [1]?: T1,
+--- [2]?: T2,
+--- [3]?: T3,
+--- [4]?: T4,
+--- [5]?: T5,
+--- [6]?: T6,
+--- [7]?: T7,
+--- [8]?: T8,
+--- [9]?: T9,
+--- [10]?: T10,
+---}
 ---@param i?   integer
 ---@param j?   integer
----@return T   ...
+---@return T1, T2, T3, T4, T5, T6, T7, T8, T9, T10
 ---@nodiscard
 function table.unpack(list, i, j) end