+++ changelog.md
@@ -1,7 +1,7 @@
 # changelog
 
 ## Unreleased
-<!-- Add all new changes here. They will be moved under a version at release -->
+* `CHG` updated Love2d version
 
 ## 3.13.2
 `2024-11-21`