+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 <!-- Add all new changes here. They will be moved under a version at release -->
+* `FIX` missing-fields diagnostic now warns about missing inherited fields
 
 ## 3.13.2
 `2024-11-21`