+++ changelog.md
@@ -2,8 +2,9 @@
 
 ## Unreleased
 <!-- Add all new changes here. They will be moved under a version at release -->
-* `FIX` missing-fields diagnostic now warns about missing inherited fields
 * `CHG` Update Love2d version
+* `CHG` Improve type infer of `table.unpack` and `unpack`
+* `FIX` missing-fields diagnostic now warns about missing inherited fields
 
 ## 3.13.2
 `2024-11-21`